requirements.txt
faster-whisper
sounddevice
scipy
numpy
keyboard
pyautogui
pyperclip
pystray
Pillow
